Chat template cleanup
