--- packages/cli/templates/pages/ai-chat-artifact/template.doc.mjs
@@ -1,11 +1,0 @@
-// Copyright (c) Meta Platforms, Inc. and affiliates.
-
-/** @type {import('../../../../core/src/docs-types').TemplateDoc} */
-export const doc = {
-  type: 'page',
-  name: 'Artifact Page',
-  displayName: 'Artifact Page',
-  description: 'AI chat interface with a resizable chat panel on the left and an artifact preview area on the right for content being created via the conversation',
-  isReady: true,
-  category: 'AI Chat - Artifact Page',
-};
