+++ packages/cli/templates/pages/ai-chat/template.doc.mjs
@@ -5,7 +5,7 @@ export const doc = {
   type: 'page',
   name: 'AI Chat Conversation',
   displayName: 'AI Chat Conversation',
-  description: 'AI assistant conversation view with tool calls, system messages, markdown, code blocks, and multi-bubble grouping',
+  description: 'AI assistant conversation view with tool calls, system messages, markdown, code blocks, multi-bubble grouping, and a resizable artifact preview panel for generated documents',
   isReady: true,
   category: 'AI Chat - Conversation',
 };
